Split queries into a separate service.
